+++ axelor-bank-payment/src/main/java/com/axelor/apps/bankpayment/service/bankorder/file/directdebit/BankOrderFile00800102Service.java
@@ -51,6 +51,7 @@ import com.axelor.apps.bankpayment.xsd.sepa.pain_008_001_02.SequenceType1Code;
 import com.axelor.apps.bankpayment.xsd.sepa.pain_008_001_02.ServiceLevel8Choice;
 import com.axelor.apps.base.db.Bank;
 import com.axelor.apps.base.db.BankDetails;
+import com.axelor.apps.tool.StringTool;
 import com.axelor.exception.AxelorException;
 import com.axelor.exception.db.repo.TraceBackRepository;
 import com.axelor.i18n.I18n;
@@ -667,7 +668,9 @@ public class BankOrderFile00800102Service extends BankOrderFile008Service {
        * Information supplied to enable the matching of an entry with the items that the transfer is intended
        * to settle, eg, commercial invoices in an accounts' receivable system in an unstructured form.
        */
-      remittanceInformation5.getUstrd().add(bankOrderLine.getReceiverReference());
+      remittanceInformation5
+          .getUstrd()
+          .add(StringTool.truncRight(bankOrderLine.getReceiverReference(), 140));
 
       /*
        * Structured   (choice 2 of 2)