+++ axelor-bank-payment/src/main/java/com/axelor/apps/bankpayment/service/bankorder/file/transfer/BankOrderFile00100102Service.java
@@ -43,6 +43,7 @@ import com.axelor.apps.bankpayment.xsd.sepa.pain_001_001_02.ServiceLevel1Code;
 import com.axelor.apps.bankpayment.xsd.sepa.pain_001_001_02.ServiceLevel2Choice;
 import com.axelor.apps.base.db.Bank;
 import com.axelor.apps.base.db.BankDetails;
+import com.axelor.apps.tool.StringTool;
 import com.axelor.exception.AxelorException;
 import com.google.common.base.Strings;
 import com.google.inject.Inject;
@@ -176,7 +177,7 @@ public class BankOrderFile00100102Service extends BankOrderFileService {
 
       String ustrd = "";
       if (!Strings.isNullOrEmpty(bankOrderLine.getReceiverReference())) {
-        ustrd += bankOrderLine.getReceiverReference();
+        ustrd += StringTool.truncRight(bankOrderLine.getReceiverReference(), 140);
       }
       if (!Strings.isNullOrEmpty(bankOrderLine.getReceiverLabel())) {
         if (!Strings.isNullOrEmpty(ustrd)) {