+++ axelor-bank-payment/src/main/java/com/axelor/apps/bankpayment/service/bankorder/file/transfer/BankOrderFile00100103Service.java
@@ -41,6 +41,7 @@ import com.axelor.apps.bankpayment.xsd.sepa.pain_001_001_03.RemittanceInformatio
 import com.axelor.apps.bankpayment.xsd.sepa.pain_001_001_03.ServiceLevel8Choice;
 import com.axelor.apps.base.db.Bank;
 import com.axelor.apps.base.db.BankDetails;
+import com.axelor.apps.tool.StringTool;
 import com.axelor.exception.AxelorException;
 import com.google.common.base.Strings;
 import com.google.inject.Inject;
@@ -174,7 +175,7 @@ public class BankOrderFile00100103Service extends BankOrderFileService {
 
       String ustrd = "";
       if (!Strings.isNullOrEmpty(bankOrderLine.getReceiverReference())) {
-        ustrd += bankOrderLine.getReceiverReference();
+        ustrd += StringTool.truncRight(bankOrderLine.getReceiverReference(), 140);
       }
       if (!Strings.isNullOrEmpty(bankOrderLine.getReceiverLabel())) {
         if (!Strings.isNullOrEmpty(ustrd)) {