+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - CUSTOMER INFORMATIONS : Indicate that Payment delay is in days
 - INVOICES DASHBOARD: Turnover is now calculated using both sales and assets
 - MANUF ORDER : fix missing form and grid view attributes for workshopStockLocation.
+- BANK ORDER LINE : add all origin references in reference field.
 
 ## Bug Fixes
 - Fix Timesheet Reminder Batch sendReminder method